[REFACTOR] : article-refactoring
article all 뷰랑 aritcle weekly 뷰랑 합침
