+++ app/src/main/java/com/indipage/data/api/ArticleApiService.kt
@@ -2,6 +2,7 @@ package com.indipage.data.api
 
 import com.indipage.data.dto.BaseResponse
 import com.indipage.data.dto.NullResponse
+import com.indipage.data.dto.response.ResponseArticleAllDto
 import com.indipage.data.dto.response.ResponseArticleSlideDto
 import com.indipage.data.dto.response.ResponseArticleWeeklyDto
 import retrofit2.http.GET
@@ -16,4 +17,7 @@ interface ArticleApiService {
 
     @PUT("/user/weekly/slide")
     suspend fun putArticleSlide(): NullResponse
+
+    @GET("/article")
+    suspend fun getArticleAll(): BaseResponse<List<ResponseArticleAllDto>>
 }
