+++ app/src/main/java/com/indipage/data/datasource/ArticleDataSource.kt
@@ -2,6 +2,7 @@ package com.indipage.data.datasource
 
 import com.indipage.data.dto.BaseResponse
 import com.indipage.data.dto.NullResponse
+import com.indipage.data.dto.response.ResponseArticleAllDto
 import com.indipage.data.dto.response.ResponseArticleSlideDto
 import com.indipage.data.dto.response.ResponseArticleWeeklyDto
 
@@ -11,4 +12,6 @@ interface ArticleDataSource {
     suspend fun getArticleSlide(): BaseResponse<ResponseArticleSlideDto>
 
     suspend fun putArticleSlide(): NullResponse
+
+    suspend fun getArticleAll(): BaseResponse<List<ResponseArticleAllDto>>
 }
