+++ app/src/main/java/com/indipage/domain/repository/ArticleDetailRepository.kt
@@ -1,6 +1,5 @@
 package com.indipage.domain.repository
 
-import com.indipage.data.dto.response.ResponseArticleAllDto
 import com.indipage.data.dto.response.ResponseArticleBookmarkDto
 import com.indipage.data.dto.response.ResponseArticleDetailDto
 import com.indipage.data.dto.response.ResponseTicketReceiveCheckDto
@@ -11,8 +10,6 @@ interface ArticleDetailRepository {
     suspend fun getTicketReceiveCheck(spaceId: Long): Result<ResponseTicketReceiveCheckDto>
 
     suspend fun postTicketReceive(spaceId: Long): Result<Int>
-
-    suspend fun getArticleAll(): Result<List<ResponseArticleAllDto>>
 
     suspend fun getBookmark(articleId: Long): Result<ResponseArticleBookmarkDto>
