+++ app/src/main/java/com/indipage/domain/repository/ArticleRepository.kt
@@ -1,5 +1,6 @@
 package com.indipage.domain.repository
 
+import com.indipage.data.dto.response.ResponseArticleAllDto
 import com.indipage.data.dto.response.ResponseArticleSlideDto
 import com.indipage.data.dto.response.ResponseArticleWeeklyDto
 
@@ -9,4 +10,6 @@ interface ArticleRepository {
     suspend fun getArticleSlide(): Result<ResponseArticleSlideDto>
 
     suspend fun putArticleSlide(): Result<Int>
+
+    suspend fun getArticleAll(): Result<List<ResponseArticleAllDto>>
 }
