+++ app/src/main/java/com/indipage/presentation/article/ArticleViewModel.kt
@@ -5,6 +5,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.core_ui.view.UiState
+import com.indipage.data.dto.response.ResponseArticleAllDto
 import com.indipage.data.dto.response.ResponseArticleSlideDto
 import com.indipage.data.dto.response.ResponseArticleWeeklyDto
 import com.indipage.domain.repository.ArticleRepository
@@ -33,9 +34,13 @@ class ArticleViewModel @Inject constructor(
     private val _openArticleDetail = MutableLiveData<Event<ResponseArticleWeeklyDto>>()
     val openArticleDetail: LiveData<Event<ResponseArticleWeeklyDto>> = _openArticleDetail
 
+    private val _articleAllData: MutableLiveData<List<ResponseArticleAllDto>> = MutableLiveData()
+    val articleAllData: LiveData<List<ResponseArticleAllDto>> = _articleAllData
+
     init {
         getArticleWeekly()
         getArticleSlide()
+        getArticleAll()
     }
 
     fun openArticleDetail(responseArticleWeeklyDto: ResponseArticleWeeklyDto) {
@@ -66,5 +71,12 @@ class ArticleViewModel @Inject constructor(
             _putArticleSlide.value = UiState.Success(400)
             Timber.d(it.message.toString())
         }
+    }
+
+    private fun getArticleAll() = viewModelScope.launch {
+        apiRepository.getArticleAll().onSuccess {
+            _articleAllData.value = it
+            Timber.d("Success")
+        }.onFailure { Timber.d(it.message.toString()) }
     }
 }
