--- app/src/main/res/layout/fragment_article_all.xml
@@ -1,60 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
-
-    <data>
-
-    </data>
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        tools:context=".presentation.article.ArticleFragment">
-
-        <androidx.appcompat.widget.AppCompatButton
-            android:id="@+id/btn_article_all_category_weekly"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="32dp"
-            android:layout_marginTop="40dp"
-            android:background="@drawable/selector_article_category_button_back"
-            android:minWidth="0dp"
-            android:minHeight="0dp"
-            android:paddingHorizontal="20dp"
-            android:paddingVertical="10dp"
-            android:text="@string/btn_article_category_weekly"
-            android:textAppearance="@style/TextAppearance.IndiPage.En.Body01"
-            android:textColor="@drawable/selector_article_category_button_text"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            tools:ignore="TouchTargetSizeCheck" />
-
-        <androidx.appcompat.widget.AppCompatButton
-            android:id="@+id/btn_article_all_category_all"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="10dp"
-            android:background="@drawable/selector_article_category_button_back"
-            android:minWidth="0dp"
-            android:minHeight="0dp"
-            android:paddingHorizontal="20dp"
-            android:paddingVertical="10dp"
-            android:text="@string/btn_article_category_all"
-            android:textAppearance="@style/TextAppearance.IndiPage.En.Body01"
-            android:textColor="@drawable/selector_article_category_button_text"
-            app:layout_constraintStart_toEndOf="@id/btn_article_all_category_weekly"
-            app:layout_constraintTop_toTopOf="@id/btn_article_all_category_weekly"
-            tools:ignore="TouchTargetSizeCheck" />
-
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/rv_article_all"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="12dp"
-            android:paddingHorizontal="32dp"
-            app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-            app:layout_constraintTop_toBottomOf="@id/btn_article_all_category_weekly"
-            tools:listitem="@layout/item_article_all" />
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</layout>
